Safeguard file deletion
This is for rare case when the trap is activated before any output to file
is generated, meaning error in internal functions

+++ lib/hook.sh
@@ -44,7 +44,7 @@ loadVars() {
 }
 
 cleanup() {
-    if [ -e "$tmpfile" ]
+    if [ -n "${tmpfile:-}" ] && [ -e "$tmpfile" ]
     then
         debug "Removing tmpfile"
         rm "$tmpfile"
